fix(api): precomputed explanations never loaded, and the failure was invisible
/api/v1/stats reported warmed: 0 while the crawler was writing 149 explanations
to disk — 93 model-written, 56 templated. The API logged the reason every thirty
seconds:

  json: cannot unmarshal number 13500.0 into Go struct field
  Explanation.rejected_numbers of type int

The guard was changed to report decimals, so rejected_numbers became a float.
Python then serializes a whole number as 13500.0, and Go's int refuses any JSON
number containing a decimal point — so the type mismatch fired on values that
were integers all along. There is not a single fractional value in the
production file; the ".0" alone was enough.

What made it hard to notice is the part worth keeping in mind. The reload path
caught the error and kept the previous set, which is the right behaviour for a
corrupt rebuild. The previous set was empty, so the correct fallback produced a
total product failure: every «چرا این؟» fell through to on-demand generation and
waited ten seconds on CPU, while the service reported itself healthy and warned
into a log nobody was reading.

Fifth Python-to-Go schema mismatch after car photos, duplicates_collapsed,
listings_captured and median_reliable — and the first that raised an error
rather than silently dropping a field. The louder failure lasted longer, because
the fallback around it worked.

Pinned with a test that parses a realistic precomputed file, decimals included.

+++ services/api/internal/index/explanations.go
@@ -15,10 +15,15 @@ import (
 // pipeline makes: cluster, normalize and now explain offline, so the request
 // path only reads.
 type Explanation struct {
-	Text            string         `json:"text"`
-	Source          string         `json:"source"`
-	Fingerprint     string         `json:"fingerprint"`
-	RejectedNumbers []int          `json:"rejected_numbers"`
+	Text        string `json:"text"`
+	Source      string `json:"source"`
+	Fingerprint string `json:"fingerprint"`
+	// float64, not int. The guard reports decimals — «۱۱.۶٪» is one number, not
+	// 116 — and an int here made every reload fail with an unmarshal error,
+	// which the reload path then handled gracefully by keeping the previous
+	// set. The previous set was empty, so the service served zero precomputed
+	// explanations indefinitely while logging a warning nobody was reading.
+	RejectedNumbers []float64      `json:"rejected_numbers"`
 	RejectedTopics  []string       `json:"rejected_topics"`
 	Usage           map[string]any `json:"usage"`
 }
